+++ src/main/java/com/nhl/link/move/runtime/json/query/Token.java
@@ -1,4 +1,4 @@
-package com.nhl.link.move.runtime.json;
+package com.nhl.link.move.runtime.json.query;
 
 class Token {
